Crate earth

rust-earth

Rust-earth is an alternative library of libearth, the shared common library for various Earth Reader apps. Earth Reader try to support many platforms as possible (e.g. web, mobile apps, desktop apps), so there must be a large part of common concepts and implementations they share like subscription lists, synchronization through cloud storages between several devices, and crawler, that libearth actually implements. Rust-earth is developing to cover the corner cases which libearth cannot be easily included like mobile devices.

Modules

codecs

feed
Data structures for feeds.

html
The adatper to display a given value as an HTML element.

mimetype

parser

repository

sanitizer

schema

stage

util