Module earth::feedUnstable

Data structures for feeds.

rust-earth internally stores archive data as Atom format, like libearth does. It's exactly not a complete set of RFC 4287, but a subset of the most of that. Since it's not intended for crawling but internal representation, it does not follow robustness principle or such thing. It simply treats stored data are all valid and well-formed.

Reexports

pub use self::category::Category;

pub use self::content::Content;

pub use self::entry::Entry;

pub use self::feed::Feed;

pub use self::generator::Generator;

pub use self::link::{Link, LinkIteratorExt, LinkList};

pub use self::mark::Mark;

pub use self::metadata::Metadata;

pub use self::person::Person;

pub use self::source::Source;

pub use self::text::Text;

Modules

category

content

entry

feed

generator

link

mark

metadata

person

source

text

Traits

Blob