Struct earth::feed::entry::EntryUnstable

pub struct Entry {
    pub metadata: Metadata,
    pub published_at: Option<DateTime<FixedOffset>>,
    pub summary: Option<Text>,
    pub content: Option<Content>,
    pub source: Option<Source>,
    pub read: Mark,
    pub starred: Mark,
}

Represent an individual entry, acting as a container for metadata and data associated with the entry. It corresponds to atom:entry element of :rfc:4287#section-4.1.2 (section 4.1.2).

Fields

metadata

published_at

The datetime value with a fixed timezone offset, indicating an instant in time associated with an event early in the life cycle of the entry. Typically, published_at will be associated with the initial creation or first availability of the resource. It corresponds to atom:published element of :rfc:4287#section-4.2.9 (section 4.2.9).

summary

The text field that conveys a short summary, abstract, or excerpt of the entry. It corresponds to atom:summary element of :rfc:4287#section-4.2.13 (section 4.2.13).

content

It either contains or links to the content of the entry. It corresponds to atom:content element of :rfc:4287#section-4.1.3 (section 4.1.3).

source

If an entry is copied from one feed into another feed, then the source feed's metadata may be preserved within the copied entry by adding source if it is not already present in the entry, and including some or all of the source feed's metadata as the source's data.

It is designed to allow the aggregation of entries from different feeds while retaining information about an entry's source feed.

It corresponds to atom:source element of :rfc:4287#section-4.2.10 (section 4.2.10).

read

Whether and when it's read or unread.

starred

Whether and when it's starred or unstarred.

Methods

impl Entry

fn new_inherited(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Entry

fn new(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Entry