Struct earth::feed::feed::FeedUnstable

pub struct Feed {
    pub source: Source,
    pub entries: Vec<Entry>,
}

Atom feed document, acting as a container for metadata and data associated with the feed.

It corresponds to atom:feed element of :rfc:4287#section-4.1.1 (section 4.1.1).

Fields

source

entries

The list of Entry values that represent an individual entry, acting as a container for metadata and data associated with the entry. It corresponds to atom:entry element of :rfc:4287#section-4.1.2 (section 4.1.2).

Methods

impl Feed

fn new_inherited(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Feed

fn new<T, S: ?Sized>(id: T, title: Text, updated_at: DateTime<FixedOffset>) -> Feed where T: Deref<Target=S>, S: ToOwned<String>