Struct earth::feed::link::LinkUnstable

pub struct Link {
    pub uri: String,
    pub relation: String,
    pub mimetype: Option<String>,
    pub language: Option<String>,
    pub title: Option<String>,
    pub byte_size: Option<u64>,
}

Link element defined in RFC 4287 (section 4.2.7).

RFC: https://tools.ietf.org/html/rfc4287#section-4.2.7.

Fields

uri

The link's required URI. It corresponds to href attribute of RFC 4287 (section 4.2.7.1).

relation

The relation type of the link. It corresponds to rel attribute of RFC 4287 (section 4.2.7.2).

See also

Existing rel values --- Microformats Wiki

This page contains tables of known HTML rel values from specifications, formats, proposals, brainstorms, and non-trivial POSH usage in the wild. In addition, dropped and rejected values are listed at the end for comprehensiveness.

mimetype

The optional hint for the MIME media type of the linked content. It corresponds to type attribute of RFC 4287 (section 4.2.7.3).

language

The language of the linked content. It corresponds to hreflang attribute of RFC 4287 (section 4.2.7.4).

title

The title of the linked resource. It corresponds to title attribute of RFC 4287 (section 4.2.7.5).

byte_size

The optional hint for the length of the linked content in octets. It corresponds to length attribute of RFC 4287 (section 4.2.7.6) .

Methods

impl Link

fn new<T, S: ?Sized>(uri: T) -> Link where T: Deref<Target=S>, S: ToOwned<String>

fn is_html(&self) -> bool

Whether its mimetype is HTML (or XHTML).