Struct earth::feed::source::SourceUnstable

pub struct Source {
    pub metadata: Metadata,
    pub subtitle: Option<Text>,
    pub generator: Option<Generator>,
    pub logo: Option<String>,
    pub icon: Option<String>,
}

All metadata for Feed excepting Feed.entries. It corresponds to atom:source element of :rfc:4287#section-4.2.10 (section 4.2.10).

Fields

metadata

subtitle

A text that conveys a human-readable description or subtitle for a feed. It corresponds to atom:subtitle element of :rfc:4287#section-4.2.12 (section 4.2.12).

generator

Identify the agent used to generate a feed, for debugging and other purposes. It corresponds to atom:generator element of :rfc:4287#section-4.2.4 (section 4.2.4).

logo

URI that identifies an image that provides visual identification for a feed. It corresponds to atom:logo element of :rfc:4287#section-4.2.8 (section 4.2.8).

icon

URI that identifies an image that provides iconic visual identification for a feed. It corresponds to atom:icon element of :rfc:4287#section-4.2.5 (section 4.2.5).

Methods

impl Source

fn new_inherited(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Source

fn new(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Source