Struct earth::feed::MetadataUnstable

pub struct Metadata {
    pub id: String,
    pub title: Text,
    pub links: Vec<Link>,
    pub updated_at: DateTime<FixedOffset>,
    pub authors: Vec<Person>,
    pub contributors: Vec<Person>,
    pub categories: Vec<Category>,
    pub rights: Option<Text>,
}

Common metadata shared by Source, Entry, and Feed.

Fields

id

The URI that conveys a permanent, universally unique identifier for an entry or feed. It corresponds to atom:id element of :rfc:4287#section-4.2.6 (section 4.2.6).

title

The human-readable title for an entry or feed. It corresponds to atom:title element of :rfc:4287#section-4.2.14 (section 4.2.14).

links

The list of :class:Link objects that define a reference from an entry or feed to a web resource. It corresponds to atom:link element of :rfc:4287#section-4.2.7 (section 4.2.7).

updated_at

The datetime value with a fixed timezone offset, indicating the most recent instant in time when the entry was modified in a way the publisher considers significant. Therefore, not all modifications necessarily result in a changed updated_at value. It corresponds to atom:updated element of :rfc:4287#section-4.2.15 (section 4.2.15).

authors

The list of Person values which indicates the author of the entry or feed. It corresponds to atom:author element of :rfc:4287#section-4.2.1 (section 4.2.1).

contributors

The list of Person values which indicates a person or other entity who contributed to the entry or feed. It corresponds to atom:contributor element of :rfc:4287#section-4.2.3 (section 4.2.3).

categories

The list of Category values that conveys information about categories associated with an entry or feed. It corresponds to atom:category element of :rfc:4287#section-4.2.2 (section 4.2.2).

rights

The text field that conveys information about rights held in and of an entry or feed. It corresponds to atom:rights element of :rfc:4287#section-4.2.10 (section 4.2.10).

Methods

impl Metadata

fn new_inherited(id: String, title: Text, updated_at: DateTime<FixedOffset>) -> Metadata