Struct earth::parser::base::XmlName

pub struct XmlName {
    pub local_name: String,
    pub namespace: Option<String>,
    pub prefix: Option<String>,
}

An owned variant of Name.

Everything about Name applies to this structure as well.

Fields

local_name

A local name, e.g. string in xsi:string.

namespace

A namespace URI, e.g. http://www.w3.org/2000/xmlns/.

prefix

A name prefix, e.g. xsi in xsi:string.

Methods

impl OwnedName

fn borrow(&self) -> Name

Constructs a borrowed Name based on this owned name.

fn local<S>(local_name: S) -> OwnedName where S: IntoOwned<String>

Returns a new OwnedName instance representing a plain local name.

fn qualified<S1, S2, S3>(local_name: S1, namespace: S2, prefix: Option<S3>) -> OwnedName where S1: IntoOwned<String>, S2: IntoOwned<String>, S3: IntoOwned<String>

Returns a new OwnedName instance representing a qualified name with or without a prefix and with a namespace URI.

fn prefix_as_ref(&self) -> Option<&str>

Returns an optional prefix by reference, equivalent to self.borrow().prefix but avoids extra work.

fn namespace_as_ref(&self) -> Option<&str>

Returns an optional namespace by reference, equivalen to self.borrow().namespace but avoids extra work.

fn to_repr(&self) -> String

See Name::to_repr() for details.