Module earth::repositoryUnstable

Abstracts storage backend e.g. filesystem.

There might be platforms that have no chance to directly access file system e.g. iOS, and in that case the concept of repository makes you to store data directly to Dropbox or Google Drive instead of filesystem. However in the most cases we will simply use FileSystemRepository even if data are synchronized using Dropbox or rsync.

Reexports

pub use self::fs::FileSystemRepository;

Modules

fs

Enums

RepositoryError

Traits

Bytes

Repository
Repository interface agnostic to its underlying storage implementation. Stage objects can deal with documents to be stored using the interface.

ToRepository

Type Definitions

Names

RepositoryResult