Trait earth::repository::RepositoryUnstable

pub trait Repository {
    fn get_reader<'a, T: BytesContainer>(&'a self, key: &[T]) -> RepositoryResult<Box<Buffer + 'a>>;
    fn get_writer<'a, T: BytesContainer>(&'a mut self, key: &[T]) -> RepositoryResult<Box<Writer + 'a>>;
    fn exists<T: BytesContainer>(&self, key: &[T]) -> bool;
    fn list<'a, T: BytesContainer>(&'a self, key: &[T]) -> RepositoryResult<Names<'a>>;

    fn read<T: BytesContainer>(&self, key: &[T]) -> RepositoryResult<Vec<u8>> { ... }
    fn write<T: BytesContainer, U: Bytes>(&mut self, key: &[T], buf: &[U]) -> RepositoryResult<()> { ... }
}

Repository interface agnostic to its underlying storage implementation. Stage objects can deal with documents to be stored using the interface.

Every content in repositories is accessible using keys. It actually abstracts out "filenames" in "file systems", hence keys share the common concepts with filenames. Keys are hierarchical, like file paths, so consists of multiple sequential strings e.g. ['dir', 'subdir', 'key']. You can list() all subkeys in the upper key as well e.g.:

repository.list(&["dir", "subdir"])

Required Methods

fn get_reader<'a, T: BytesContainer>(&'a self, key: &[T]) -> RepositoryResult<Box<Buffer + 'a>>

Read the content from the key.

fn get_writer<'a, T: BytesContainer>(&'a mut self, key: &[T]) -> RepositoryResult<Box<Writer + 'a>>

Get a writer to write data into the key.

fn exists<T: BytesContainer>(&self, key: &[T]) -> bool

Return whether the key exists or not.

fn list<'a, T: BytesContainer>(&'a self, key: &[T]) -> RepositoryResult<Names<'a>>

List all subkeys in the key.

Provided Methods

fn read<T: BytesContainer>(&self, key: &[T]) -> RepositoryResult<Vec<u8>>

fn write<T: BytesContainer, U: Bytes>(&mut self, key: &[T], buf: &[U]) -> RepositoryResult<()>