Trait earth::repository::ToRepositoryUnstable

pub trait ToRepository<R: Repository> {
    fn to_repo(&self) -> RepositoryResult<R>;
    fn from_repo(repo: &R, scheme: &str) -> Self;
}

Required Methods

fn to_repo(&self) -> RepositoryResult<R>

Create a new instance of the repository from itself. It may be used for configuring the repository in plain text e.g. *.ini.

fn from_repo(repo: &R, scheme: &str) -> Self

Generate a value that to_repo() can accept. It's used for configuring the repository in plain text e.g. *.ini. URL scheme is determined by caller, and given through argument.