Module earth::sanitizerExperimental

Sanitize HTML tags.

Structs

CleanHtml

Escape

SanitizeHtml

Functions

clean_html
Strip all markup tags from HTML string. That means, it simply makes the given HTML document a plain text.

escape
Convert given string to HTML-safe sequences by replacing the characters &, < and >. If the optional flag quote is true, the characters " and ' are also translated.

sanitize_html
Sanitize the given HTML string. It removes the following tags and attributes that are not secure nor useful for RSS reader layout: