Trait earth::schema::MergeableExperimental

pub trait Mergeable: Sized {
    fn merge_with(&mut self, other: Self) { ... }
}

This trait is used to merge conflicts between concurrent updates.

Note

The default implementation does nothing. That means the entity of the newer session will always win unless the method is redefined.

Provided Methods

fn merge_with(&mut self, other: Self)

Merge data with the given value to renew itself as a latest state.

The given argument other is guaranteed that it's from the older session. (note that it doesn't mean this entity is older than self, but the last update of the session is)