Module earth::feedUnstable

Data structures for feeds.

rust-earth internally stores archive data as Atom format, like libearth does. It's exactly not a complete set of RFC 4287, but a subset of the most of that. Since it's not intended for crawling but internal representation, it does not follow robustness principle or such thing. It simply treats stored data are all valid and well-formed.

Structs

Category
Category element defined in :rfc:4287#section-4.2.2 (section 4.2.2).

Content
Content construct defined in :rfc:4287#section-4.1.3 (section 4.1.3).

Entry
Represent an individual entry, acting as a container for metadata and data associated with the entry. It corresponds to atom:entry element of :rfc:4287#section-4.1.2 (section 4.1.2).

Feed
Atom feed document, acting as a container for metadata and data associated with the feed.

Generator
Identify the agent used to generate a feed, for debugging and other purposes. It's corresponds to atom:generator element of :rfc:4287#section-4.2.4 (section 4.2.4).

Link
Link element defined in RFC 4287 (section 4.2.7).

LinkList

Mark
Represent whether the entry is read, starred, or tagged by user.

Metadata
Common metadata shared by Source, Entry, and Feed.

Person
Person construct defined in RFC 4287 (section 3.2).

Source
All metadata for Feed excepting Feed.entries. It corresponds to atom:source element of :rfc:4287#section-4.2.10 (section 4.2.10).

Enums

Text
Text construct defined in :rfc:4287#section-3.1 (section 3.1).

Traits

Blob

LinkIteratorExt